cleanup, remove duplicates
