+++ packages/entrykit/src/icons/ArrowLeftIcon.tsx
@@ -2,7 +2,7 @@ import { IconSVG, Props } from "./IconSVG";
 
 export function ArrowLeftIcon(props: Props) {
   return (
-    <IconSVG strokeWidth="2" stroke="currentColor" {...props}>
+    <IconSVG {...props}>
       <path
         d="M19 12H5M12 19l-7-7 7-7"
         fill="none"
