+++ packages/entrykit/src/onboarding/deposit/ChainBalance.tsx
@@ -1,7 +1,7 @@
 import { Hex } from "viem";
 import { useBalance } from "wagmi";
 import { PendingIcon } from "../../icons/PendingIcon";
-import { Balance } from "./Balance";
+import { Balance } from "../../ui/Balance";
 
 export type Props = {
   chainId: number;
@@ -10,5 +10,5 @@ export type Props = {
 
 export function ChainBalance({ chainId, address }: Props) {
   const balance = useBalance({ chainId, address });
-  return <>{balance.data ? <Balance amount={balance.data.value} /> : <PendingIcon />}</>;
+  return <>{balance.data ? <Balance wei={balance.data.value} /> : <PendingIcon />}</>;
 }
