+++ packages/entrykit/src/onboarding/deposit/RelayDepositStatus.tsx
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { DepositStatus } from "./DepositStatus";
 import { useChains } from "wagmi";
 import { RelayDeposit } from "./useDeposits";
-import { Balance } from "./Balance";
+import { Balance } from "../../ui/Balance";
 
 export type Props = RelayDeposit & { onDismiss: () => void };
 
@@ -45,7 +45,7 @@ export function RelayDepositStatus({
         }
         return (
           <>
-            Successfully bridged <Balance amount={amount} /> to {chainL2.name}!
+            Successfully bridged <Balance wei={amount} /> to {chainL2.name}!
           </>
         );
       })()}
