--- packages/entrykit/src/useAppChain.tsx
@@ -1,16 +1,0 @@
-import { useChains } from "wagmi";
-import { MUDChain } from "@latticexyz/common/chains";
-import { useEntryKitConfig } from "./EntryKitConfigProvider";
-
-export function useAppChain(): MUDChain {
-  const { chainId } = useEntryKitConfig();
-  const chains = useChains();
-  const chain = chains.find((c) => c.id === chainId);
-
-  if (!chain) {
-    throw new Error(
-      `Account Kit is configured to use chain ${chainId}, but the corresponding chain was not found in wagmi's configured chains.`,
-    );
-  }
-  return chain;
-}
